Crate small_bytecode_interpreter

Structs

Interpreter

Enums

ByteCode

(1) You are a TA at a university, and you want to evaluate your student’s homework without executing their (untrusted) code. You decide to write a small web-service that takes bytecode as input, and interprets the results. The bytecode language you need to support includes basic arithmetic and variables. The bytecode language is stack, rather than register based. ByteCode (right) is given for the following pseudo code (left):